Frame switching W3C Compliance
This change updates the frame functionality to comply with the W3C
specification.

Valid values of the `id` parameter are:
- null
- a Number between 0 and 2^16
- an Object representing a Web element

When the value is null, set the current browsing context to the current
top-level browsing context

When the value is a valid Number, let window be the associated window of
the current browsing context’s active document.

When the value is an Object, let element be the result of trying to get
a known element by web element reference id.

Source: https://www.w3.org/TR/webdriver/#switch-to-frame