Make the capability conversion JsonSerializable
By making the ChromeOptions implement JsonSerializable, both the
ChromeOptions::CAPABILITY and ChromeOptions::CAPABILITY_W3C will be
serializable when converting to JSON for the Desired Capabilities.

This also solves an issue whereby it is not possible to specify
CAPABILITY_W3C as the capability key as merging then works correctly
following this change.

Fixes #849